Publish script

+++ deploy.sh
@@ -1,3 +1,28 @@
+#!/bin/sh
+
+if [ "`git status -s`" ]
+then
+    echo "The working directory is dirty. Please commit any pending changes."
+    exit 1;
+fi
+
+echo "Deleting old publication"
 rm -rf build
+mkdir build
+git worktree prune
+rm -rf .git/worktrees/build/
+
+echo "Checking out gh-pages branch into build"
+git worktree add -B gh-pages build origin/gh-pages
+
+echo "Removing existing files"
+rm -rf build/*
+
+echo "Generating site"
 hugo -D
-cd build && git add --all && git commit -m "Publishing to gh-pages" && git push && cd ..
+
+echo "Updating gh-pages branch"
+cd build && git add --all && git commit -m "Publishing to gh-pages (publish.sh)"
+
+#echo "Pushing to github"
+git push --all
